requirements.txt
six
cffi
django
djangorestframework
django-rest-auth
django-allauth
django-cors-headers
markdown
django-filter
pyyaml
more_itertools
pynacl
django-redis
ntplib
python-dateutil
django-storages
dropbox
azure
apache-libcloud
pyblake2
